+++ Makefile
@@ -6,7 +6,7 @@
 #    By: twang <twang@student.42.fr>                +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/03/20 14:09:46 by twang             #+#    #+#              #
-#    Updated: 2023/05/17 14:43:07 by twang            ###   ########.fr        #
+#    Updated: 2023/05/17 15:30:04 by twang            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -37,7 +37,7 @@ OBJ_DIR		=	.objs
 
 #--flags-----------------------------------------------------------------------#
 
-CFLAGS		=	-g3 -Wall -Wextra -I $(LIBFT_DIR) -I $(INC_DIR)/execution_incs -I $(INC_DIR)/parsing_incs -o2 #-Werror
+CFLAGS		=	-g3 -Wall -Wextra -I $(LIBFT_DIR) -I $(INC_DIR) -I $(INC_DIR) #-Werror
 
 #--debug flags--------------------------------------------------------#
