--- config/headers_execution.mk
@@ -1,20 +1,0 @@
-# **************************************************************************** #
-#                                                                              #
-#                                                         :::      ::::::::    #
-#    headers_execution.mk                               :+:      :+:    :+:    #
-#                                                     +:+ +:+         +:+      #
-#    By: twang <twang@student.42.fr>                +#+  +:+       +#+         #
-#                                                 +#+#+#+#+#+   +#+            #
-#    Created: 2023/03/20 13:32:02 by twang             #+#    #+#              #
-#    Updated: 2023/05/09 12:36:46 by twang            ###   ########.fr        #
-#                                                                              #
-# **************************************************************************** #
-
-HEADERS	=	incs/execution_incs/defines_execution.h							\
-			incs/execution_incs/enums_execution.h							\
-			incs/execution_incs/minishell_execution.h						\
-			incs/execution_incs/prototypes_execution.h						\
-			incs/execution_incs/structures_execution.h						\
-			libft/libft.h 													\
-			srcs/debug_folder/debug.h
-
