--- config/headers_parsing.mk
@@ -1,18 +1,0 @@
-# **************************************************************************** #
-#                                                                              #
-#                                                         :::      ::::::::    #
-#    headers_parsing.mk                                 :+:      :+:    :+:    #
-#                                                     +:+ +:+         +:+      #
-#    By: twang <twang@student.42.fr>                +#+  +:+       +#+         #
-#                                                 +#+#+#+#+#+   +#+            #
-#    Created: 2023/03/20 13:32:02 by twang             #+#    #+#              #
-#    Updated: 2023/03/27 16:35:41 by twang            ###   ########.fr        #
-#                                                                              #
-# **************************************************************************** #
-
-HEADERS	+=	incs/parsing_incs/defines_parsing.h								\
-			incs/parsing_incs/enums_parsing.h								\
-			incs/parsing_incs/minishell_parsing.h							\
-			incs/parsing_incs/prototypes_parsing.h							\
-			incs/parsing_incs/structures_parsing.h
-
