--- config/sources_execution.mk
@@ -1,32 +1,0 @@
-# **************************************************************************** #
-#                                                                              #
-#                                                         :::      ::::::::    #
-#    sources_execution.mk                               :+:      :+:    :+:    #
-#                                                     +:+ +:+         +:+      #
-#    By: twang <twang@student.42.fr>                +#+  +:+       +#+         #
-#                                                 +#+#+#+#+#+   +#+            #
-#    Created: 2023/03/20 13:31:59 by twang             #+#    #+#              #
-#    Updated: 2023/05/16 17:54:41 by twang            ###   ########.fr        #
-#                                                                              #
-# **************************************************************************** #
-
-SOURCES	=																	\
-			srcs/bonus/wildcards/files_wildcards.c							\
-			srcs/bonus/wildcards/wildcards_utils.c							\
-			srcs/bonus/wildcards/wildcards.c								\
-			srcs/execution/execve/childs.c									\
-			srcs/execution/execve/fds_manage.c								\
-			srcs/execution/parsing/clean_utils.c							\
-			srcs/execution/parsing/heredoc_manage.c							\
-			srcs/execution/parsing/infiles_manage.c							\
-			srcs/execution/parsing/outfiles_manage.c						\
-			srcs/execution/parsing/strings_manage.c							\
-			srcs/execution/utils/cmds_utils.c								\
-			srcs/execution/utils/files_utils.c								\
-			srcs/execution/utils/path_utils.c								\
-			srcs/execution/utils/lstc_utils.c								\
-			srcs/execution/utils/struct_utils.c								\
-			srcs/execution/execution.c										\
-			srcs/signals/signals.c											\
-			srcs/prompt.c													\
-			srcs/debug_folder/debug_fonction.c
