--- incs/parsing_incs/defines_parsing.h
@@ -1,21 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   defines_parsing.h                                  :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: twang <twang@student.42.fr>                +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/03/20 18:27:48 by wangthea          #+#    #+#             */
-/*   Updated: 2023/05/17 10:33:29 by twang            ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-/*#ifndef DEFINES_PARSING_H
-# define DEFINES_PARSING_H
-
-# define MINI_SHELL_MUST_GO_ON 1
-# define STRT_CHR -1
-# define EOSTR 0
-# define METACHAR_SET "|&<>() \t\v\f\r\n"
-# define IFS " \t\n"
-#endif */
